Remove unused --current-jar option
This was added in the initial drop of Metalava so it is not clear
whether it was actually used or why it was added but it appears to have
been to allow the current version of the code for have provided by
`prebuilts/sdk/current/public/android.jar`. Irrespective of that it is
no longer used, not even in a test.

Given that we are working towards generating `api-versions.xml` from
signature files even if we did wish to support providing the current
API from something other than sources then we would need to provide
more signature files.

Either way it is not used and will not be used so this change removes
it.

Bug: 378479241
Test: ./gradlew
Change-Id: Iba0743b52f7ee5a2a4afbeff286346a989e46bf6