Remove unnecessary setting of global options
Previously, the `ReportingOptions.reporter` (now called
`IssueReportingOptions.bootstrapReporter`) was a `DefaultReporter`
which directly accessed the global `options` property when an issue
was reported. That required `options` to be initialized in a test that
checked the behavior when an unknown issues was provided that would
cause the reporter to report an issue.

However, that is no longer true so the initialization of `options` is
no longer needed in the test so this change removes it.

Bug: 295013652
Test: ./gradlew
Change-Id: I6bffad5c7848b7b1a3839368305e5300c337d17e