Delegate Options.currentApiLevel
Previously, the `currentApiLevel` property in `Options` was built upon
the `currentApiLevel` in `ApiLevelsGenerationOptions`. This change
renames the latter to `optionalCurrentApiLevel` and makes it private
and then copies the `Options` implementation of `currentApiLevel`. The
existing `Options` property were rewritten to delegate to the new one.

This is part of a series of changes to move functionality needed by
`ApiGenerator.generateXml(...)` into `ApiLevelsGenerationOptions` so
it can abstract away a lot of the details.

Bug: 378479241
Test: ./gradlew
Change-Id: I1b50417a911388979f150322008628dcb06939ba