Use numeric API level in Javadoc if --current-version not provided
Previously, if `--current-codename` was provided by `--current-version`
was not then the codename would be used for all the API levels. That
was happening because when `--current-version` was not specified it
would use the codename when the level was greater than the default of
`-1`. As API levels are greater than or equal to 1 that was always.

This change switches the default to `Int.MAX_VALUE`. That meant that
all callers of `currentApiLevelOrDefault(int)` were passing in
`Int.MAX_VALUE` as the `default` parameter so it removed the default
and changed it to a property.

Bug: 378479241
Test: ./gradlew
Change-Id: I01531d71649698b96e4e587938bafcbee0d9f232