Test providing --current-codename but not --current-version
While working on cleaning up `DocAnalyzer` an edge case was discovered
where if `--current-codename` was specified but `--current-version` was
not then the codename would be used to represent all API levels. This
change adds a test to cover that which shows the incorrect behavior.

This does not appear to cause problems for Metalava users which means
they must provide both or neither. However, fixing it will allow some
inconsistencies in the handling of this to be removed.

Bug: 378479241
Test: ./gradlew
Change-Id: Id43a0152ea458ea29e11b04952bb97604efecc56