Use `null` for the unset value of ApiElement.deprecatedIn
Previously, `ApiElement.deprecatedIn` used `0` as a special value to
indicate that it was not set. That made it difficult to reason about
as it was no clear whether it was checking the unset value. This change
makes it nullable so it can use `null` as the unset value. That makes
it explicit when the unset value is being used.

Also removed the `ApiElement.isDeprecated` property as it was only
used internally and provided no real value.

Bug: 379646963
Test: ./gradlew
Change-Id: I98e74d8bcb5d48bd6e0fcd67ccaa1ecd3b59e2bd