Replicate Jvm annotation failure in K2
The @JvmName and @JvmStatic annotations aren't being picked up in K2 with this project structure. The issue seems to be around the `root dir` property in the project xml. Removing it makes the test pass for K2.

Similar to other K2 issues we've seen where androidx annotations aren't picked up.

Bug: 360195094
Bug: 370786649
Test: adds test (that fails with K2)

Change-Id: I100da9bc8f7a83256e3f8745aac538cc4581c104