Simplify handling of SDK ids in DocAnalyzer
Previously, `DocAnalyzer` would keep track of information about the
version of the `ANDROID_PLATFORM_SDK_ID` (0) in which a symbol was
added, all the way from reading the `api-versions.xml` file up to
writing out the `@sdkExtSince` doctag where it would discard it.

The `id`/`version` would first be stored as `first`/`second` in
`IdAndVersion` and then as `sdk`/`version` in `SdkAndVersion`. It
would be filtered out when selecting the `SdkAndVersion` that is
written out for `@sdkExtSince`.

This change filters the `id`/`version` where `id` is 0, i.e. the
Android Platform SDK, out when reading the `sdks` attribute, avoiding
the need to do it later. It also allows the `sdk` property to be
removed from `SdkAndVersion` and removes the entry for
`ANDROID_PLATFORM_SDK_ID` in `sdkIdentifiers` as it no longer needs
to map from the id to the name (which is stored in `SdkAndVersion`).

Bug: 380412656
Test: ./gradlew
Change-Id: I9e8e2852d481f6ad3113a203cad4c78cb7b0ce49