Abstract ApiElement updating into an Updater class
Previously, when adding APIs from a jar or `Codebase` to an `Api`
object the `SdkVersion` would be passed through the code that processed
the jar of `Codebase` and eventually passed to the
`ApiElement.update(...)` method. This wraps the `SdkVersion` into an
`Updater` class and passes an instance of that instead.

That will allow a follow-up change to pass a specialized `Updater`
instance that will support updating an `ApiElement` for an SDK
extension allowing special SDK extension handling to be removed from
`Api.readJar(...)` and making it easier in future to support loading
SDK extension information from signature files instead of jar files.

Bug: 379646963
Test: ./gradlew
Change-Id: Idcc924f5bb6c0938e245f4f8b454161decd76a50