Remove references to JSON from generateJson(...)
Previously, `generateJson(...)` was differentiated from
`generateXml(...)` by two factors, it consumed signature files and it
generated a JSON file. The latter is no longer true so this change
corrects the name of the method, the configuration and other uses of
`json` to make them match the implementation.

This is separated from the previous change to make them both easier to
review.

Bug: 380412656
Test: ./gradlew
Change-Id: Ib5f8278ec55c25480fe29971d222c76f7d28761a