Fix computation of version in removed attribute
The `ApiXmlPrinter` determines that if an `ApiElement`'s parent
appeared in a later version than the `ApiElement` then the `ApiElement`
was removed in the version following its last appearance. It assumed
that that version could be computed simply by adding one to the version
in which it last appeared. That was true when the version was simply a
single `Int` and the versions where a contiguous range on integers.
However, it is not true for more complicated versions such as ones with
separate `major` and `minor` components. Nevermind ones as complicated
as those used in `AndroidX`.

This change fixes that issue by passing in a list of all versions, in
order and then constructing a map from each version to the next one in
the list. That fixed an issue with an existing test.

Bug: 380412656
Test: ./gradlew
Change-Id: Ib6d8cb0fc9a0e287a024e02ed47f48e70c6aa0b8