Add test for removing classes in current codebase
The `ApiXmlPrinter` marks a class as removed if
`ApiClass.lastPresentIn` (the last version the class was present in) is
less than `Api.lastPresentIn` (the last version of the API being
tracked). That indicates that the class was not in the last version of
the API. However, the version of the current codebase is not recorded
in `Api.lastPresentIn` and so if a class is `@removed` in the current
codebase it is not treated as having been removed in the
generated file.

e.g. if version 1-35 are available in prebuilts and 36 is the version
of the current `Codebase` then if a class is `@removed` in the current
source then `ApiClass.lastPresentIn` would be set to `35` and
`Api.lastPresentIn` would also be set to `35`, so the class is not
treated as having been removed. However, if the `Api.lastPresentIn` was
correctly set to `36` then the class would be treated as having been
removed.

This change adds a test to show the incorrect behavior. A follow-up
change will fix it.

Bug: 379646963
Test: ./gradlew
Change-Id: Ic6e36ca5046a27d725a4455866aacef605b99c14