Stop passing version/deprecated to ApiElement constructor
Previously, `ApiElement` constructor would take an `SdkVersion` and a
`deprecated` flag and would initialize its properties with them. It
would also then duplicate some of that behavior in the `update(...)`
method.

This stops passing version/deprecated into the constructor and instead
defers the initialization of the properties to the `update(...)`
method. This does mean that an `ApiElement` is not fully initialized
until after the `update(...)` method is called but it does mean that
there is only one way to set those properties. That will simplify
follow-up changes.

Bug: 379646963
Test: ./gradlew
Change-Id: I6c8da72719026833319ddbd7c18e4851bdb2e9ad