Add ApiElement.Updater for SDK extensions
Previously, `Api.readJar(...)` had special support for updating
`ApiElement`s for an SDK extension version. This change adds a
`ExtensionUpdater` that abstracts away the code for updating an
`ApiElement` for an SDK extension and removes the special support in
`Api.readJar(...)`.

This should make it easier in future to load SDK extension information
from signature files instead of jar files.

Bug: 379646963
Test: ./gradlew
Change-Id: I7e42885cc2216c5caa39185aa9106a247ec69c83