Encapsulate SignatureFileLoader in VersionedSignatureApi
Previously, the `VersionedSignatureApi` was passed the
`SignatureFileLoader` in its `load(...)` method. That required it to be
passed into the `ApiGenerator`. This change encapsulates the
`SignatureFileLoader` in the `VersionedSignatureApi`. This is
preparation for a follow-up change that will move the functionality to
update the `Api` into `VersionedSignatureApi` so that an abstraction
can be created which will update an `Api` from either signature files
or `.jar` files.

Bug: 378479241
Test: ./gradlew
Change-Id: Ie23ab87deba0bcef66c1e963141ac7d95ada2cb0