Move API updating into VersionedSignatureApi
Previously, `ApiGenerator.createApiFromSignatureFiles(...)` would get a
`Codebase` and `ApiVersion` from a `VersionedSignatureApi` and use them
to update the `Api`. This moves the updating into an `updateApi(Api)`
method on `VersionedSignatureApi`.

This will allow follow-up changes to create an abstraction for a
`VersionedApi` that can have different implementations, e.g. from a
signature file or the current `Codebase` or a `.jar` file.

Bug: 378479241
Test: ./gradlew
Change-Id: I84a26c0b6fc3b323628ef120bccf0f3ab497aaf0