Replace GenerateXmlConfig.apiLevels with versionToJar Map
Previously, `apiLevels` was a `List<File>` where the index corresponded
to the API level. Only those entries whose index was greater than or
equal to `firstApiLevel` were valid; invalid entries had an invalid
`File`. That implicit mapping complicated code that used it.

This change replaces it with a map from `ApiVersion` to `File`,
simplifying the code that uses it. It also removed the
`GenerateXmlConfig.firstApiLevel` property as it was no longer needed.

Bug: 378479241
Test: ./gradlew
Change-Id: Id625a82c73f8b4b3ef1072f0386c778fdc45949e