Fix failure when trying to finalize an SDK release
Previously, any extension APIs added since the last finalized version
were assumed to be added to the current version + 1. While that was ok
for developer builds it was not ok for finalization builds, i.e. builds
creating the finalized artifacts for the next release as it lead to
extension APIs being added in current version + 1 which broke the
`ApiXmlPrinter` logic for determining if an API had been removed and
that in turn lead to an exception being thrown.

This change fixes that and improves the error message that is thrown by
`ApiXmlPrinter` giving more information for future issues.

Bug: 383521289
Test: ./gradlew
      # Testing on Android
Change-Id: I3918d505c56e314d73bda3242a4710b038c8957f