Make CreateAggregateLibraryBuildInfoFileTask to be Isolated Project compatible
Move to a pattern of publishing configuration on each project that adds
build info as an artifact and then a consuming configuration on the root
project that will collect all of these build info files to aggregate
them. This allows root and other projects to be configured concurrently
and is Gradle Isolated Project feature safe.

This pattern is explained in https://www.liutikas.net/2024/12/11/Together-In-Isolation.html

Bug: None
Test: ./gradlew
Change-Id: I1c58ec3c095fa706650740ec0aff5d5fc020879b