Update ApiLint acronym search
The hasAcronyms and getFirstAcronym methods duplicated some logic for searching for acronyms, so this updates hasAcronym to use getFirstAcronym. It also updates the acronym search to use one pattern for all acronym lengths that matches only the acronym itself, not the start of the word. This will simplify the followup change, which will allow certain acronyms.

Bug: 350519983
Test: AcronymNamesTest
Change-Id: I233956108df08022fe9fba970633c3211e878be0