Implicitly propagate @RequiresOptIn annotations on properties to setters and getters
The Kotlin compiler treats @RequiresOptIn annotations on properties to also include setters and getters. Additionally it is unsupported by the compiler to explicitly annotate the getter, so the only valid way to mark a getter on a property is by marking the property itself. This updates Metalava to match the Kotlin compiler behavior.

Bug: 174708311
Test: ApiFileTest
Change-Id: I1be44e3ab49592a1bb908b4e6c00f5bf8df71382