Make isHiddenOrRemoved() call hidden() instead of hidden
Previously, it called `hidden` which meant that it would not treat an
item as hidden if the containing class was hidden. Fixing it caused
changes in signature files in Android. This change fixes it but adds a
legacy implementation that is used for checking fields to maintain the
existing behavior.

That will be fixed in follow-up changes once downstream has been fixed.

Bug: 300211291
Test: ./gradlew
Change-Id: I4d1a7e52627b9cdcc399dbd1018a38f6b787909a